requirements.txt
asgiref==3.3.4
boto3==1.17.62
botocore==1.20.62
CacheControl==0.12.6
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
Django==3.2
djangorestframework==3.12.4
firebase-admin==5.0.0
gcloud==0.17.0
google-api-core==1.28.0
google-api-python-client==2.6.0
google-auth==1.30.1
google-auth-httplib2==0.1.0
google-cloud-core==1.6.0
google-cloud-firestore==2.1.1
google-cloud-storage==1.38.0
google-crc32c==1.1.2
google-resumable-media==1.3.0
googleapis-common-protos==1.53.0
grpcio==1.38.0
httplib2==0.19.1
idna==2.10
jmespath==0.10.0
jws==0.1.3
msgpack==1.0.2
oauth2client==3.0.0
packaging==20.9
pip==21.1.2
proto-plus==1.18.1
protobuf==3.16.0rc1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.4.3
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.17.1
python-jwt==2.0.1
pytz==2021.1
requests==2.25.1
requests-toolbelt==0.7.0
rsa==4.7.2
s3transfer==0.4.2
setuptools==49.6.0.post20210108
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.4
wheel==0.36.2
wincertstore==0.2
